Make commands wrap.
Many commands in vim stop at line endings (or beginnings) by default.
This is not the case with moders editors, where a <Right> at the end of
a line will put you at the beginning of the next line. Being as this
matches the current users intuitive grasp of how text editors work, it
is worth making vim do the same.