Add function to modify timestamps of files.
When supplied a path, the Set-FileTimeStamps function will iterate over
the child objects and update the modified time, created time and
accessed time to be the current date (or a supplied date) as if they
were created at that time and untouched since then.  This is similar to
the unix `touch` command.