Changes fish_prompt to be more extensible.
Alteres the way fish_prompt functions to be more extensible - instead of
echoing each section as it is generated, they are saved to variables
local to the function.  At the end, these are converted into an array
and iterated through to output.  This will allow for adding/removing
sections more easily in the future and should make it possible to have
sections be displayed more "smartly" when the terminal windows is small.
(This is a planned future feature.)