Make scripts more portable.
Scripts now use the `/usr/bin/env` shebang trick, to account for
different locations of executables.  Although this is technically an
abuse of the `env` utility, it is widely used, and most distros are
unlikely to do anything to break it.  Usage on older systems will
probably require modifications to the script anyway.