Change rules for git commits.
We now use the more sensible autocmd on entering a file of the
appropriate type (as opposed to checking the filename) and have dropped
the colorscheme. Instead of forcing me to remember I was editing
a commit, I simply found it jarring.  I committed something, I know I'm
editing a commit message!