Change plugin manager from Vundle to Vim-Plug.
Vim-Plug is a more focused plugin than Vundle, and it offers the benefit
of async plugin installation/upgrades, which is nice on new systems or
systems that haven't been used for a while.

The install script has also been altered, and there is a windows version
added, which should allow quick and easy installation on all systems.