Enable lazy-loading.
For all plugins where it makes sense, added settings to make vim-plug
load them only when needed (on a specific command, or a particular
file type). After this, the plugin will still be active globally, but
this does trade off startup time for a smaller (and presumably less
noticeable) wait later.