Successfully deletes group.
Still have "backwards typing" issue.

I suspect `getcurpos()`, `line('.')` and friends jump to Ex-mode, and when
they return to Insert mode, they press i instead of a (or something
similar?). I'm not comfortable enough with (n)vim internals to comment
on this.

I will read the docs more, and maybe the source.  If anyone besides me
is following this, don't hold your breath. It's more a proof of concept
at this point than an actual project.