Allow more room on screen when viewing diffs.
When viewing diffs, most of the time there is already a shortage of
horizontal space. Further, the most common use case is to be comparing
commits from a repositories history, which have (hopefully) already been
linted. In any case, it is rare that you will need ALE to be active in
diff mode, and it is also rare for diff mode to lead into editing
directly, so disabling ALE has few, if any, negative side effects in
this instance.