Reformat ftplugins.
The line continuations in vimscript are sort of hacky and fragile. (Look
at the comment in the python.vim file. Having it right after the
`setlocal` made that line a bare invocation, so it just listed my local
settings. Then every line after that was silently ignored, so my
settings weren't even applied, to boot!).

The new form is slightly less elegent but it's readable and less
error-prone.