Remove vim-plug guards.
vim-plug can only load plugins in specific cases, such as when editing
a specific filetype or when a specific command is called. This is
intended to reduce startup time. I've done a bit of testing and found
that to be a negligible gain, and so the added complexity isn't needed.