Add more flexibility to `tm` command.
Now will attach to extant sessions easily, if it can find a match for
that session.

This can cause collisions, so don't name a session "ls", I guess. Or,
use the alternate form, `tm attach -t ls`, which will still work - if
the number of arguments is greater than one, it will simply forwatd the
command to tmux.